refactor: Separate scraping logic into bridge service
Refactored the application to a two-service architecture. A new 'bridge' service now encapsulates all Playwright browser automation, isolating it from the main 'addon' service. This improves stability, scalability, and resource management.

The 'addon' service communicates with the 'bridge' service via an internal HTTP API, as orchestrated by Docker Compose. The core scraping logic was moved from 'lib/handlers.js' into the new 'bridge-server.js'. A security warning was also added to the README to advise against exposing the bridge service publicly.
