+++ lib/handlers.js
@@ -7,90 +7,11 @@ const {
     normalizeId
 } = require("./scraping");
 
-const NodeCache = require('node-cache');
-const streamCache = new NodeCache({ stdTTL: 3600, checkperiod: 600 });
+// All browser-related logic has been moved to the bridge server.
 
-const PROVIDERS = {
-    'yourupload.com': async (page) => {
-        await page.waitForSelector('video');
-        return page.evaluate(() => document.querySelector('video')?.src || document.querySelector('video source')?.src);
-    },
-    'mp4upload.com': async (page) => {
-        await page.waitForSelector('video', { state: 'visible', timeout: 20000 });
-        return page.evaluate(() => {
-            const scripts = Array.from(document.querySelectorAll('script'));
-            for (const script of scripts) {
-                const match = script.textContent.match(/https?:\/\/[^"']+\.(mp4|m3u8)[^"']*/);
-                if (match) return match[0];
-            }
-            const video = document.querySelector('video');
-            return video?.src || video?.querySelector('source')?.src;
-        });
-    },
-    'vidsrc.to': async (page) => {
-        await page.waitForSelector('iframe');
-        const iframeSrc = await page.$eval('iframe', el => el.src);
-        if (iframeSrc.includes('m3u8')) return iframeSrc;
-        return page.evaluate(() => {
-            const scripts = document.querySelectorAll('script');
-            for (const script of scripts) {
-                const match = script.textContent.match(/https?:\/\/[^"']+\.m3u8[^"']*/);
-                if (match) return match[0];
-            }
-            return null;
-        });
-    }
-};
-
-async function extractVideoUrl(browser, url, referer = null) {
-    const cacheKey = `video_url:${url}`;
-    const cached = streamCache.get(cacheKey);
-    if (cached) {
-        console.log(`Cache hit for final URL: ${url}`);
-        return cached;
-    }
-
-    const page = await browser.newPage();
-    try {
-        await page.setExtraHTTPHeaders({
-            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
-            'Referer': referer || new URL(url).origin,
-        });
-        await page.goto(url, { waitUntil: 'networkidle', timeout: 30000 });
-
-        const detectedProvider = Object.keys(PROVIDERS).find(p => url.includes(p));
-        const extractor = detectedProvider ? PROVIDERS[detectedProvider] : null;
-
-        let videoUrl;
-        if (extractor) {
-            videoUrl = await extractor(page);
-        } else {
-            videoUrl = await page.evaluate(() => {
-                const video = document.querySelector('video');
-                if (video?.src) return video.src;
-                const source = document.querySelector('video source');
-                if (source?.src) return source.src;
-                const scripts = Array.from(document.querySelectorAll('script'));
-                for (const script of scripts) {
-                    const match = script.textContent.match(/https?:\/\/[^\s"']+\.(?:m3u8|mp4)[^\s"']*/);
-                    if (match) return match[0];
-                }
-                return null;
-            });
-        }
-
-        if (videoUrl) {
-            streamCache.set(cacheKey, videoUrl);
-        }
-        return videoUrl;
-    } finally {
-        await page.close();
-    }
-}
-
-function defineHandlers(builder, getBrowser) {
+function defineHandlers(builder) {
     builder.defineCatalogHandler(async ({type, id, extra}) => {
-        console.log("request for catalog: " + type + " " + id);
+        console.log("[Addon] request for catalog: " + type + " " + id);
 
         let url;
         if (extra && extra.search) {
@@ -345,14 +266,14 @@ function defineHandlers(builder, getBrowser) {
     }
 
     builder.defineStreamHandler(async ({ type, id }) => {
-        console.log("request for streams: " + type + " " + id);
+        console.log("[Addon] request for streams: " + type + " " + id);
         let targetUrl;
 
         if (id.startsWith('tt')) {
             const [imdbId, season, episode] = id.split(':');
             targetUrl = await getLatanimeEpisodeUrl(imdbId, parseInt(season), parseInt(episode));
             if (!targetUrl) {
-                console.error(`Could not resolve IMDb ID ${id} to a Latanime URL.`);
+                console.error(`[Addon] Could not resolve IMDb ID ${id} to a Latanime URL.`);
                 return Promise.resolve({ streams: [] });
             }
         } else {
@@ -360,105 +281,22 @@ function defineHandlers(builder, getBrowser) {
             targetUrl = `${LATANIME_URL}/ver/${episodeId}`;
         }
 
-        const browser = getBrowser();
-        if (!browser) {
-            console.error('Browser is not ready');
-            return Promise.resolve({ streams: [] });
-        }
+        const bridgeUrl = process.env.BRIDGE_URL || 'http://bridge:3001';
+        console.log(`[Addon] Requesting streams from bridge: ${bridgeUrl} for target: ${targetUrl}`);
 
-        console.log(`Scraping latanime page: ${targetUrl}`);
-        const page = await browser.newPage();
         try {
-            await page.goto(targetUrl, { waitUntil: 'domcontentloaded' });
-
-            const providers = await page.evaluate(() => {
-                const results = [];
-                const baseKey = document.querySelector('div.player')?.getAttribute('data-key');
-                if (!baseKey) return [];
-                const basePlayerUrl = atob(baseKey);
-                document.querySelectorAll('a.play-video').forEach(el => {
-                    const providerName = el.textContent.trim();
-                    const encodedPart = el.getAttribute('data-player');
-                    if (encodedPart) {
-                        const intermediateUrl = providerName.toLowerCase() === 'yourupload' ? atob(encodedPart) : basePlayerUrl + encodedPart;
-                        results.push({ url: intermediateUrl, title: providerName });
-                    }
-                });
-                return results;
-            });
-            console.log(`Found ${providers.length} potential providers.`);
-
-            const resolvedStreams = [];
-
-            for (const provider of providers) {
-                const providerPage = await browser.newPage();
-                try {
-                    // Optimization: Don't load images or CSS
-                    await providerPage.setRequestInterception(true);
-                    providerPage.on('request', (req) => {
-                        if (['image', 'stylesheet', 'font'].includes(req.resourceType())) {
-                            req.abort();
-                        } else {
-                            req.continue();
-                        }
-                    });
-
-                    await providerPage.goto(provider.url, { waitUntil: 'domcontentloaded', timeout: 15000 });
-
-                    const finalUrl = await providerPage.evaluate(() => {
-                        const redirMatch = document.body.innerHTML.match(/var redir = atob\("([^"]+)"\);/);
-                        return redirMatch ? atob(redirMatch[1]) : null;
-                    });
-
-                    if (finalUrl && !finalUrl.includes('listeamed.net')) {
-                        console.log(`Found valid final embed URL for ${provider.title}: ${finalUrl.substring(0, 60)}...`);
-                        const videoUrl = await extractVideoUrl(browser, finalUrl, provider.url);
-                        if (videoUrl) {
-                            console.log(`✅ Extracted: ${provider.title} -> ${videoUrl.substring(0, 60)}...`);
-                            resolvedStreams.push({
-                                name: 'Latanime',
-                                url: videoUrl,
-                                title: provider.title,
-                                behaviorHints: {
-                                    proxyHeaders: {
-                                        'request': {
-                                            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36',
-                                            'Referer': new URL(finalUrl).origin
-                                        }
-                                    }
-                                }
-                            });
-                        }
-                    }
-                } catch (e) {
-                    console.error(`Failed processing provider ${provider.title} (${provider.url}): ${e.message}`);
-                } finally {
-                    await providerPage.close();
-                }
-            }
-
-            const downloadLinks = await page.evaluate(() => {
-                const links = [];
-                const selectors = [
-                    'a[href*="pixeldrain.com"]', 'a[href*="mediafire.com"]', 'a[href*="mega.nz"]',
-                    'a[href*="gofile.io"]', 'a[href*="drive.google.com"]', 'a[href*="1fichier.com"]',
-                    'a[download]'
-                ];
-                document.querySelectorAll(selectors.join(',')).forEach(el => {
-                    if (el.href) links.push({ url: el.href, title: `📥 ${el.textContent.trim() || 'Download'}` });
-                });
-                return links;
-            });
-
-            const allStreams = [...resolvedStreams, ...downloadLinks];
-            console.log(`Total streams found: ${allStreams.length}`);
-            return Promise.resolve({ streams: allStreams });
-
+            // The bridge service is now responsible for the entire scraping process.
+            const response = await axios.post(`${bridgeUrl}/scrape`,
+                { url: targetUrl },
+                { timeout: 90000 } // 90-second timeout for the entire scraping and extraction process
+            );
+            return Promise.resolve({ streams: response.data.streams || [] });
         } catch (error) {
-            console.error(`Scraping error on ${targetUrl}: ${error.message}`);
+            console.error(`[Addon] Error fetching streams from bridge for ${targetUrl}:`, error.message);
+            if (error.response) {
+                console.error('[Addon] Bridge Error Response:', error.response.data);
+            }
             return Promise.resolve({ streams: [] });
-        } finally {
-            await page.close();
         }
     });
 }
